refactor: use crate from deno repo
