+++ Cargo.toml
@@ -2,9 +2,6 @@
 resolver = "2"
 members = ["rs_lib"]
 
-[workspace.dependencies]
-sys_traits = "0.1.16"
-
 [profile.release]
 codegen-units = 1
 incremental = true
