+++ deno_dir_test.ts
@@ -1,4 +1,4 @@
-// Copyright 2018-2025 the Deno authors. MIT license.
+// Copyright the Deno authors. MIT license.
 
 import { assertEquals, assertThrows } from "@std/assert";
 import { DenoDir } from "./deno_dir.ts";
