+++ deps_test.ts
@@ -1,4 +1,4 @@
-// Copyright 2018-2025 the Deno authors. MIT license.
+// Copyright the Deno authors. MIT license.
 
 export { assertEquals, assertRejects } from "@std/assert";
 export { createGraph } from "@deno/graph";
