+++ http_cache.ts
@@ -1,12 +1,11 @@
-// Copyright 2018-2025 the Deno authors. MIT license.
+// Copyright the Deno authors. MIT license.
 
 import { isAbsolute } from "@std/path";
 import { assert } from "./util.ts";
 import {
-  type GlobalHttpCache,
-  instantiate,
-  type LocalHttpCache,
-} from "./lib/deno_cache_dir.generated.js";
+  GlobalHttpCache,
+  LocalHttpCache,
+} from "./lib/deno_cache_dir_wasm.js";
 
 export interface HttpCacheCreateOptions {
   root: string;
@@ -47,8 +46,6 @@ export class HttpCache implements Disposable {
         "Vendor root must be an absolute path.",
       );
     }
-    const { GlobalHttpCache, LocalHttpCache } = await instantiate();
-
     let cache: LocalHttpCache | GlobalHttpCache;
     if (options.vendorRoot != null) {
       cache = LocalHttpCache.new(
