+++ rs_lib/Cargo.toml
@@ -1,34 +1,21 @@
 [package]
-name = "deno_cache_dir"
-version = "0.27.0"
+name = "deno_cache_dir_wasm"
+version = "0.0.0"
 edition = "2024"
 license = "MIT"
-description = "Cache directory logic used in Deno"
-repository = "https://github.com/denoland/deno_cache_dir"
 
 [lib]
 crate-type = ["cdylib", "lib"]
-
-[features]
-default = ["file_fetcher"]
-file_fetcher = ["async-trait", "base64", "cache_control", "chrono", "data-url", "http"]
-wasm = ["console_error_panic_hook", "js-sys", "serde-wasm-bindgen", "wasm-bindgen", "sys_traits/wasm"]
-sync = []
+path = "./lib.rs"
 
 [dependencies]
-async-trait = { version = "0.1.73", optional = true }
 base32 = "=0.5.1"
-base64 = { version = "0.21.7", optional = true }
 boxed_error = "0.2.3"
-cache_control = { version = "0.2.0", optional = true }
-# Note: Do not use the "clock" feature of chrono, as it links us to CoreFoundation on macOS.
-chrono = { version = "0.4", default-features = false, features = ["std"], optional = true }
-data-url = { version = "0.3.0", optional = true }
+deno_cache_dir = "0.27.0"
 deno_error = { version = "0.7.0", features =["url"] }
 deno_media_type = "0.4.0"
 deno_path_util = "0.6.0"
-sys_traits.workspace = true
-http = { version = "1", optional = true }
+sys_traits = { version = "0.1.24", features = ["wasm"] }
 indexmap = { version = "2.0.0", features = ["serde"] }
 log = "0.4.19"
 once_cell = "1.18.0"
@@ -39,13 +26,7 @@ sha2 = "0.10.0"
 thiserror = "2"
 url = { version = "2.5.1", features = ["serde"] }
 
-console_error_panic_hook = { version = "0.1.6", optional = true }
-js-sys = { version = "=0.3.68", optional = true }
-wasm-bindgen = { version = "=0.2.91", optional = true }
-serde-wasm-bindgen = { version = "0.6.5", optional = true }
-
-[dev-dependencies]
-pretty_assertions = "1.4.0"
-sys_traits = { workspace = true, features = ["memory", "real", "getrandom", "libc", "winapi"] }
-tempfile = "3.7.1"
-tokio = { version = "1", features = ["rt", "macros"] }
+console_error_panic_hook = "0.1.7"
+js-sys = "=0.3.83"
+wasm-bindgen = "=0.2.106"
+serde-wasm-bindgen = "0.6.5"
