+++ rust-toolchain.toml
@@ -1,3 +1,3 @@
 [toolchain]
-channel = "1.91.1"
+channel = "1.93.1"
 components = ["clippy", "rustfmt"]
